refactor: rename to Query.ts
